+++ product_service/src/models/schemas/product/update_product_input.py
@@ -6,4 +6,4 @@ class UpdateProductInput(BaseModel):
     name: str | None = None
     description: str | None = None
     price: float | None = None
-    stock: int | None = None
+    stock: float | None = None
